requirements.txt
# Core dependencies
langchain>=0.1.8
docling>=0.2.3
openai>=1.24.0
fpdf==1.7.2
langchain-openai>=0.0.1

# Document processing
pytesseract>=0.3.10
Pillow>=10.0.1
pypdf>=3.17.1
python-docx>=1.0.1
markdown>=3.5
python-pptx>=0.6.21

# OCR support
pdfminer.six>=20221105
pdf2image>=1.16.3

# Data handling
numpy>=1.25.2
pandas>=2.1.1

# Vector stores and embeddings
sentence-transformers>=2.2.2
qdrant-client>=1.7.0

# Utility packages
tqdm>=4.66.1
requests>=2.31.0
python-dotenv>=1.0.0
shortuuid>=1.0.11 